# spec/oapi-spec.yaml
openapi: "3.0.0"
info:
  version: "0.0.1"
  title: "OnPicket API"
  description: ""
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: http://localhost:9898/api
paths:
  /healthz:
    get:
      summary: Healthcheck endpoint for the API
      operationId: healthz
      responses:
        '200':
          description: Health endpoint
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Healthz'
  /scans:
    get:
      summary: Return all scans
      operationId: listScans
      responses:
        '200':
          description: Scan list response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scan'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new Scan
      operationId: createScan
      requestBody:
        $ref: '#/components/requestBodies/ScanBody'
      responses:
        '201':
          description: Scan response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '405':
          description: Method not found
        '400':
          description: Bad request

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


components:
  schemas:
    Scan:
      required:
        - host
        - ports
#        - description
      properties:
        host:
          type: string
        ports:
          type: array
    NewScan:
      required:
        - host
        - ports
      properties:
        host:
          type: string
        ports:
          type: array

    Error:
      type: object
      required:
        - code
        - status
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        status:
          type: string
          description: Error message
        body:
          type: object

    Healthz:
      type: object
      required:
        - status
        - version
      properties:
        status:
          type: string
        version:
          type: string
  requestBodies:
    ScanBody:
      description: JSON of new Scan
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NewScan'